requirements.txt
# ── Quantum ────────────────────────────────────────────────────────────────────
pennylane==0.32.0

# ── Classical ML ───────────────────────────────────────────────────────────────
torch==2.3.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# ── Image & data processing ────────────────────────────────────────────────────
Pillow==10.0.0
opencv-python==4.8.0.76
matplotlib==3.7.2
seaborn==0.12.2

# ── Dashboard ──────────────────────────────────────────────────────────────────
streamlit==1.35.0

# ── Utilities ──────────────────────────────────────────────────────────────────
tqdm==4.66.4
pyyaml==6.0.1
pytest==8.2.0
